Several new updates:
1. Moved INSTALL file to inst folder
2. Added COPYING file to inst folder
3. New arguments: 
   AdjWidth for approximate auto column width adjustments to longest entry in column
   AutoFilter for set up autofiltering for each column
   BoldHeaderRow to add bold font to header row entries
   FreezeRow and FreezeCol to set up frozen panes in each worksheet
4. Minor fixes to WriteXLS.Rd
5. Updated copyrights to 2010
6. Bumping to version 1.9.0

Note that after researching the potential for being able to append new worksheets to an existing XLS file, this does not appear to be a robust option. The Perl package Spreadsheet::ParseExcel does not support the preservation of many worksheet objects as noted in http://search.cpan.org/~jmcnamara/Spreadsheet-WriteExcel-2.37/lib/Spreadsheet/WriteExcel.pm#MODIFYING_AND_REWRITING_EXCEL_FILES. Thus, given this limitation using Perl and the potential for corrupting existing XLS files, there are no plans to add the ability to append new worksheets to an existing file to this package.


git-svn-id: file:///Users/marcschwartz/Downloads/VersionControl/git/R-Forge/WriteXLS/pkg@23 798c95a1-3a3c-4699-83bb-5843777bddb4