Added support to write 0 row data frames to the Excel files. Bumping to version 3.5.1.
Per request, added the ability to write 0 row data frames to the Excel file worksheet. This would result in a single header row in the sheet. This was lost with version 3.0.0 due to the rbind()ing of the comments to the data frame. With a 0 row data frame, the original column names are lost after the rbind()ing. Bumping to version 3.5.1.